Rename DiAC to AC
Split them into the 2 different libraries
