+++ QGL/PulsePrimitives/__init__.py
@@ -1,2 +1,2 @@
 from .common_primitives import Id, Z, X90, X90m, Y90, Y90m, Z90, Z90m, Utheta, Xtheta, Ytheta, Ztheta, U90, U, arb_axis_drag, MEAS, BLANK
-from .composite_primitives import AC, DiAC, CNOT, CNOT_CR, flat_top_gaussian, echoCR, ZX90_CR, MeasEcho, X, Y
+from .composite_primitives import AC, CNOT, CNOT_CR, flat_top_gaussian, echoCR, ZX90_CR, MeasEcho, X, Y, Xm, Ym
