+++ QGL/PulsePrimitives/composite_primitives.py
@@ -26,257 +26,11 @@ from .helpers import overrideDefaults, _memoize, clear_pulse_cache
 from .common_primitives import Id, Z, X90, X90m, Y90, Y90m, Z90, Z90m, Utheta, Xtheta, Ytheta, Ztheta, U90, U, arb_axis_drag
 from .. import config
 if config.pulse_primitives_lib == 'standard':
-    from .standard_primitives import X, Xm, Y, Ym
+    from .standard_primitives import X, Xm, Y, Ym, AC
 elif config.pulse_primitives_lib == 'all90':
-    from .all90_primitives import X, Xm, Y, Ym
+    from .all90_primitives import X, Xm, Y, Ym, AC
 else:
     raise Exception("Invalid pulse library")
-
-def AC(qubit, cliffNum):
-    """
-
-    The set of 24 Atomic Clifford single qubit pulses.
-
-    Parameters
-    ----------
-    qubit : logical channel to implement sequence (LogicalChannel)
-    cliffNum : the zero-indexed Clifford number
-
-    Returns
-    -------
-    pulse object
-    """
-
-    #Figure out the approximate nutation frequency calibration from the X180 and the samplingRate
-    Xp = X(qubit)
-    xpulse = Xp.amp * Xp.shape
-    nutFreq = 0.5 / (sum(xpulse) / qubit.physChan.samplingRate)
-
-    #Now a big else if chain for to get the specific Clifford
-    if cliffNum == 0:
-        #Identity gate
-        return Id(qubit, length=0)
-    elif cliffNum == 1:
-        #X90
-        return X90(qubit)
-    elif cliffNum == 2:
-        #X180
-        return X(qubit)
-    elif cliffNum == 3:
-        #X90m
-        return X90m(qubit)
-    elif cliffNum == 4:
-        #Y90
-        return Y90(qubit)
-    elif cliffNum == 5:
-        #Y180
-        return Y(qubit)
-    elif cliffNum == 6:
-        #Y90m
-        return Y90m(qubit)
-    elif cliffNum == 7:
-        #Z90
-        return Z90(qubit)
-    elif cliffNum == 8:
-        #Z180
-        return Z(qubit)
-    elif cliffNum == 9:
-        #Z90m
-        return Z90m(qubit)
-    elif cliffNum == 10:
-        #X+Y 180
-        return U(qubit, phase=pi / 4, label="AC_10")
-    elif cliffNum == 11:
-        #X-Y 180
-        return U(qubit, phase=-pi / 4, label="AC_11")
-    elif cliffNum == 12:
-        #X+Z 180(Hadamard)
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=pi,
-                             polarAngle=pi / 4,
-                             label="AC_12")
-    elif cliffNum == 13:
-        #X-Z 180
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=pi,
-                             polarAngle=pi / 4,
-                             aziAngle=pi,
-                             label="AC_13")
-    elif cliffNum == 14:
-        #Y+Z 180
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=pi,
-                             polarAngle=pi / 4,
-                             aziAngle=pi / 2,
-                             label="AC_14")
-    elif cliffNum == 15:
-        #Y-Z 180
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=pi,
-                             polarAngle=pi / 4,
-                             aziAngle=-pi / 2,
-                             label="AC_15")
-    elif cliffNum == 16:
-        #X+Y+Z 120
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=2 * pi / 3,
-                             polarAngle=acos(1 / sqrt(3)),
-                             aziAngle=pi / 4,
-                             label="AC_16")
-    elif cliffNum == 17:
-        #X+Y+Z -120 (equivalent to -X-Y-Z 120)
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=2 * pi / 3,
-                             polarAngle=pi - acos(1 / sqrt(3)),
-                             aziAngle=5 * pi / 4,
-                             label="AC_17")
-    elif cliffNum == 18:
-        #X-Y+Z 120
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=2 * pi / 3,
-                             polarAngle=acos(1 / sqrt(3)),
-                             aziAngle=-pi / 4,
-                             label="AC_18")
-    elif cliffNum == 19:
-        #X-Y+Z 120 (equivalent to -X+Y-Z -120)
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=2 * pi / 3,
-                             polarAngle=pi - acos(1 / sqrt(3)),
-                             aziAngle=3 * pi / 4,
-                             label="AC_19")
-    elif cliffNum == 20:
-        #X+Y-Z 120
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=2 * pi / 3,
-                             polarAngle=pi - acos(1 / sqrt(3)),
-                             aziAngle=pi / 4,
-                             label="AC_20")
-    elif cliffNum == 21:
-        #X+Y-Z -120 (equivalent to -X-Y+Z 120)
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=2 * pi / 3,
-                             polarAngle=acos(1 / sqrt(3)),
-                             aziAngle=5 * pi / 4,
-                             label="AC_21")
-    elif cliffNum == 22:
-        #-X+Y+Z 120
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=2 * pi / 3,
-                             polarAngle=acos(1 / sqrt(3)),
-                             aziAngle=3 * pi / 4,
-                             label="AC_22")
-    elif cliffNum == 23:
-        #-X+Y+Z -120 (equivalent to X-Y-Z 120)
-        return arb_axis_drag(qubit,
-                             nutFreq,
-                             rotAngle=2 * pi / 3,
-                             polarAngle=pi - acos(1 / sqrt(3)),
-                             aziAngle=-pi / 4,
-                             label="AC_23")
-    else:
-        raise ValueError('Clifford number must be between 0 and 23')
-
-
-def DiAC(qubit, cliffNum, Xonly = False):
-    """
-
-    The set of 24 Diatomic Clifford single qubit pulses.
-
-    Parameters
-    ----------
-    qubit : logical channel to implement sequence (LogicalChannel)
-    cliffNum : the zero-indexed Clifford number
-
-    Returns
-    -------
-    pulse object
-    """
-    #Now a big else if chain for to get the specific Clifford
-    if cliffNum == 0:
-        #Identity gate
-        return Id(qubit, length=0)
-    elif cliffNum == 1:
-        #X90
-        return X90(qubit)
-    elif cliffNum == 2:
-        #X180
-        return X90(qubit)+X90(qubit)
-    elif cliffNum == 3:
-        #X90m
-        return X90m(qubit)
-    elif cliffNum == 4:
-        #Y90
-        return X90(qubit) + Z90m(qubit) + X90(qubit) + Z(qubit) if Xonly else Y90(qubit)
-    elif cliffNum == 5:
-        #Y180
-        return X90(qubit) + X90(qubit) + Z(qubit) if Xonly else Y90(qubit) + Y90(qubit)
-    elif cliffNum == 6:
-        #Y90m
-        return X90(qubit) + Z90(qubit) + X90(qubit) + Z(qubit) if Xonly else Y90m(qubit)
-    elif cliffNum == 7:
-        #Z90
-        return Z90(qubit)
-    elif cliffNum == 8:
-        #Z180
-        return Z(qubit)
-    elif cliffNum == 9:
-        #Z90m
-        return Z90m(qubit)
-    elif cliffNum == 10:
-        #X+Y 180
-        return X90(qubit) + X90(qubit) + Z90(qubit) if Xonly else Y90(qubit) + Y90(qubit) + Z90m(qubit)
-    elif cliffNum == 11:
-        #X-Y 180
-        return X90(qubit) + X90(qubit) + Z90m(qubit) if Xonly else Y90(qubit) + Y90(qubit) + Z90(qubit)
-    elif cliffNum == 12:
-        #X+Z 180(Hadamard)
-        return Z(qubit) + X90(qubit) + Z90m(qubit) + X90(qubit) + Z(qubit) if Xonly else Z(qubit) + Y90(qubit)
-    elif cliffNum == 13:
-        #X-Z 180
-        return Z(qubit) + X90(qubit) + Z90(qubit) + X90(qubit) + Z(qubit) if Xonly else Z(qubit) + Y90m(qubit)
-    elif cliffNum == 14:
-        #Y+Z 180
-        return Z90(qubit) + X90(qubit) + Z90m(qubit) + X90(qubit) + Z90m(qubit) if Xonly else Z90(qubit) + Y90(qubit) + Z90(qubit)
-    elif cliffNum == 15:
-        #Y-Z 180
-        return Z90(qubit) + X90(qubit) + Z90(qubit) + X90(qubit) + Z90m(qubit) if Xonly else Z90(qubit) + Y90m(qubit) + Z90(qubit)
-    elif cliffNum == 16:
-            #X+Y+Z -120 (equivalent to -X-Y-Z 120)
-        return Z90(qubit) + X90(qubit) + Z90m(qubit) + X90(qubit) + Z(qubit) if Xonly else Z90(qubit) + Y90(qubit)
-    elif cliffNum == 17:
-        #X+Y+Z 120
-        return Z(qubit) + X90(qubit) + Z90m(qubit) + X90(qubit) + Z90m(qubit) if Xonly else Z(qubit) + Y90(qubit) + Z90(qubit)
-    elif cliffNum == 18:
-            #X-Y+Z 120 (equivalent to -X+Y-Z 120)
-            return X90(qubit) + Z90(qubit) + X90(qubit) + Z90m(qubit) if Xonly else Y90m(qubit) + Z90(qubit)
-    elif cliffNum == 19:
-        #X-Y+Z -120
-        return Z90m(qubit) + X90(qubit) + Z90m(qubit) + X90(qubit) + Z(qubit) if Xonly else Z90m(qubit) + Y90(qubit)
-    elif cliffNum == 20:
-            #X+Y-Z -120 (equivalent to -X-Y+Z 120)
-            return Z(qubit) + X90(qubit) + Z90(qubit) + X90(qubit) + Z90m(qubit) if Xonly else Z(qubit) + Y90m(qubit) + Z90(qubit)
-    elif cliffNum == 21:
-        #X+Y-Z 120
-        return Z90(qubit) + X90(qubit) + Z90(qubit) + X90(qubit) + Z(qubit) if Xonly else Z90(qubit) + Y90m(qubit)
-    elif cliffNum == 22:
-        #-X+Y+Z -120 (equivalent to X-Y-Z 120)
-        return X90(qubit) + Z90m(qubit) + X90(qubit) + Z90m(qubit) if Xonly else Y90(qubit) + Z90(qubit)
-    elif cliffNum == 23:
-        #-X+Y+Z 120
-        return Z90m(qubit) + X90(qubit) + Z90(qubit) + X90(qubit) + Z(qubit) if Xonly else Z90m(qubit) + Y90m(qubit)
-    else:
-        raise ValueError('Clifford number must be between 0 and 23')
 
 ## two-qubit primitivies
 def CNOT(source, target, **kwargs):
